Rewrite procedure application code
The current implementation applies procedures by substituting
placeholders (Parameter objects), in the combination with the
arguments passed in to the procedure call.  Simplify the code by just
setting a local environment that has the passed arguments in its
namespace.  This enables reuse of the evaluation code and removes
unnecessary substitution code and the Parameter class.