Add code and tests for new Expression class
Currently eval() in Environment uses lists to represent expressions.
To avoid the not so nice calls to isinstance to determine if the list
represents a combination, create this new object to make the code
clearer, i.e. use a member method called is_combination().  The next
step is to replace the lists with instances of this Expression class.