Add support for tail recursion
In eval(), if called by a procedure that is tail recursive, raise a
TailRecursion exception when it is about to evaluate the calling
procedure again.  The exception is caught in the procedure, which
calls eval() with the arguments given to it by the TailRecursion
instance.

This will prevent the Python interpreter from nesting deeper and
deeper in recursive calls, where it eventually will hit the maximum
recursion depth.