Add BuiltinProcedure class and refactor tests
Add BuiltinProcedure which is a subclass of Procedure, and move the
functionality for calling functions implemented in Python, from
Procedure into BuiltinProcedure.  This will make it more clear that
there are two different types of procedures and that they work
slightly different.  Also refactor procedure tests to make them
clearer.